+++ src/test/promisifyEvent.js
@@ -1,5 +1,7 @@
 'use strict';
 
+const { AssertionError } = require('assert');
+
 /**
  * Promisifies an event.
  * @param {object} emitter An event emitter with `on` and `off` methods.
@@ -11,9 +13,17 @@
  */
 module.exports = function promisifyEvent(emitter, event, timeout = 1000) {
   return new Promise((resolve, reject) => {
+    // Ensure the timeout error stack trace starts at the location where
+    // `promisifyEvent` is called. Creating the error within the `setTimeout`
+    // callback function would result in an unhelpful stack trace.
+    const timeoutError = new AssertionError({
+      message: `Event “${event}” ${timeout} millisecond timeout.`,
+      stackStartFn: promisifyEvent,
+    });
+
     const timer = setTimeout(() => {
       emitter.off(event, listener);
-      reject(new Error(`Event “${event}” wait timeout.`));
+      reject(timeoutError);
     }, timeout);
 
     /**